refactorizations. TODO: add proper references to dependency kb-core, removing duplication of classes
