--- app/it/almawave/linkeddata/kb/repo/RDFRepositoryException.scala
@@ -1,3 +1,0 @@
-package it.almawave.linkeddata.kb.repo
-
-class RDFRepositoryException(message: String, cause: Throwable) extends RuntimeException
