--- app/it/almawave/linkeddata/kb/utils/README.md
@@ -1,6 +1,0 @@
-
-SEE: for better Try {} handling
-
-https://stackoverflow.com/questions/20632250/how-can-i-express-finally-equivalent-for-a-scalas-try
-
-// SEE: https://stackoverflow.com/questions/20632250/how-can-i-express-finally-equivalent-for-a-scalas-try
